hnsw: add ability to configure ef search
I needed a quick solution for using HNSW in Go and found this package
nice, but noticed that it forces `ef` = `efConstruction`. I'm optimizing
for a use case where I build an effectively static index. Build
time doesn't matter to me but query speed and recall do. So I want to be
able to set `efConstruction` high but keep `ef` where it is.

I've left this behaving in exactly the same manner as before as the
default.