Don't drop collaborative filter cache
This code will always delete collaborative filter scores for all users that are older than the ones created just now for the current user.

This effectively means that the cache will only ever have collaborative filter results for a single user.

I'm assuming the intention here was to only purge stale scores for the user that was just updated.